changed warn + strike
