+++ src/action/buildCommands.ts
@@ -12,38 +12,12 @@ export default [
   /*
     Mod-Commands
     */
-  new SlashCommandBuilder().setName('warn')
-    .setDefaultMemberPermissions(PermissionFlagsBits.KickMembers)
-    .setDescription('Verwarnt eine Person')
-    .addUserOption(
-      opt => opt.setName('target')
-        .setDescription('Die Person, die verwarnt werden soll')
-        .setRequired(true)
-    )
-    .addStringOption(
-      opt => opt.setName('reason')
-        .setDescription('Der Grund für den /warn')
-        .setRequired(true)
-    ),
   new SlashCommandBuilder().setName('rename')
     .setDefaultMemberPermissions(PermissionFlagsBits.KickMembers)
     .setDescription('Ändert deinen Benutzernamen in einen random Namen auf dem Server')
     .addUserOption(
       opt => opt.setName('target')
         .setDescription('Die Person, dessen Name geändert werden soll')
-        .setRequired(true)
-    ),
-  new SlashCommandBuilder().setName('strike')
-    .setDefaultMemberPermissions(PermissionFlagsBits.KickMembers)
-    .setDescription('Verwarnt eine Person und erteilt einen Strike.')
-    .addUserOption(
-      opt => opt.setName('target')
-        .setDescription('Die Person, die einen Strike bekommen soll')
-        .setRequired(true)
-    )
-    .addStringOption(
-      opt => opt.setName('reason')
-        .setDescription('Der Grund für den /strike')
         .setRequired(true)
     ),
   new SlashCommandBuilder().setName('timeout')
